pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semantica"
version = "1.0.0"
description = "Semantic search engine tool"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "faiss-cpu",
    "fastembed",
    "spacy",
    "termcolor",
    "pypdf",
    "python-docx",
    "psutil"
]

[project.scripts]
semantica = "semantica.main:main"

[tool.setuptools.packages.find]
where = ["src"]